Update screen Update to new Arduino framework Update to M5Unified library Use peripheral for rotary encoder Use WDT Use IPv6 Improve furnace failover

+++ src/encoder.h
@@ -2,5 +2,5 @@
 
 #include <Arduino.h>
 
-void    init_encoder();
-int16_t get_encoder();
+void init_encoder();
+int get_encoder();
